ajustes a los daos de bebida,sopa,principio,proteína. Creación de test dao para los mencionados con anterioridad.

+++ Fuentes/ServiceVentaComida/src/main/java/co/edu/usbcali/dao/IBebidaDAO.java
@@ -1,5 +1,6 @@
 package co.edu.usbcali.dao;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import co.edu.usbcali.modelo.Bebida;
@@ -9,7 +10,7 @@ public interface IBebidaDAO {
 	public void crear(Bebida entity);
 	public void modificar(Bebida entity);
 	public void borrar(Bebida entity);
-	public Bebida consultarPorId(int idBebida);
+	public Bebida consultarPorId(BigDecimal idBebida);
 	public Bebida consultarPorIdentificacion(Integer idBebida);
 	public List<Bebida> consultarTodos();
