+++ Fuentes/ServiceVentaComida/src/main/java/co/edu/usbcali/dao/IPrincipioDAO.java
@@ -1,5 +1,6 @@
 package co.edu.usbcali.dao;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import co.edu.usbcali.modelo.Principio;
@@ -9,7 +10,7 @@ public interface IPrincipioDAO {
 	public void crear(Principio entity);
 	public void modificar(Principio entity);
 	public void borrar(Principio entity);
-	public Principio consultarPorId(int idPrincipio);
+	public Principio consultarPorId(BigDecimal idPrincipio);
 	public Principio consultarPorIdentificacion(Integer idPrincipio);
 	public List<Principio> consultarTodos();
