+++ Fuentes/ServiceVentaComida/src/main/java/co/edu/usbcali/dao/IProteinaDAO.java
@@ -1,5 +1,6 @@
 package co.edu.usbcali.dao;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import co.edu.usbcali.modelo.Proteina;
@@ -9,7 +10,7 @@ public interface IProteinaDAO {
 	public void crear(Proteina entity);
 	public void modificar(Proteina entity);
 	public void borrar(Proteina entity);
-	public Proteina consultarPorId(int idProteina);
+	public Proteina consultarPorId(BigDecimal idProteina);
 	public Proteina consultarPorIdentificacion(Integer idProteina);
 	public List<Proteina> consultarTodos();
