+++ Fuentes/ServiceVentaComida/src/main/java/co/edu/usbcali/dao/ISopaDAO.java
@@ -1,5 +1,6 @@
 package co.edu.usbcali.dao;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import co.edu.usbcali.modelo.Sopa;
@@ -9,7 +10,7 @@ public interface ISopaDAO {
 	public void crear(Sopa entity);
 	public void modificar(Sopa entity);
 	public void borrar(Sopa entity);
-	public Sopa consultarPorId(int idSopa);
+	public Sopa consultarPorId(BigDecimal idSopa);
 	public Sopa consultarPorIdentificacion(Integer idSopa);
 	public List<Sopa> consultarTodos();
